% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.r
\name{const_t}
\alias{const_t}
\title{Definition function: 63\% time constant after Aubinet (2012) Eq. 4.22}
\usage{
const_t(f0)
}
\arguments{
\item{Currently}{none}
}
\value{
Currently none
}
\description{
63\% time constant after Aubinet (2012) Eq. 4.22.
}
\examples{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\references{
Currently none
}
\seealso{
Currently none
}
\keyword{Currently}
\keyword{none}

