% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.r
\name{COSP.fwd}
\alias{COSP.fwd}
\title{Definition function: Generate cospectra}
\usage{
COSP.fwd(FS, TS, cparho = 1)
}
\arguments{
\item{Currently}{none}
}
\value{
Currently none
}
\description{
Generate cospectra.
}
\examples{
Currently none
}
\references{
Currently none
}
\seealso{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\keyword{Currently}
\keyword{none}
