% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.r
\name{COSP.plot}
\alias{COSP.plot}
\title{Definition function: Plot cospectra}
\usage{
COSP.plot(name, IDE, DEP, title, labx = NA, laby = NA, colvec = c(1,
  2, 1), limx = NULL, limy = NULL, weight = T, wl = F,
  norm = NULL, part_cov = 1, modspec = T, zeta, FX = NULL,
  devran = NULL, plotting = T)
}
\arguments{
\item{Currently}{none}
}
\value{
Currently none
}
\description{
Plot cospectra.
}
\examples{
Currently none
}
\references{
Currently none
}
\seealso{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\keyword{Currently}
\keyword{none}
