% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.dist.isca.R
\name{def.dist.isca}
\alias{def.dist.isca}
\title{Definition function: Integral length scales}
\usage{
def.dist.isca(scalEddy, data, veloXaxs = NULL)
}
\arguments{
\item{\code{scalEddy}}{A vector containing distances or times and of class "numeric". [m] or [s]}

\item{\code{data}}{A vector containing the input data. Of class "numeric" or "integer". [user-defined]}

\item{\code{veloXaxs}}{Instantaneous wind speed. Only supplied when users define scalEddy in terms of time. If provided, of class "numeric", otherwise NULL. [m/s]}
}
\value{
Estimated Integral length scale.
}
\description{
Function defintion. Integral length scales.
}
\examples{
Currently none
}
\references{
J. Bange, F. Beyrich, D.A.M. Engelbart (2002) Airbourne measurements of turbulent fluxes during LITFASS-98: Comparison with ground measurements and remote sensing in a case study. Theor. Appl. Climatol., 73, 35-51. \cr 
D.H. Lenschow, B.B.Stankov. (1986) Length Scales in the Convective Boundary Layer. Journal of the Atmospheric Sciences. 43:12, 1198-1209 \cr
}
\seealso{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
Kenny Pratt
}
\keyword{eddy-covariance,}
\keyword{flux}
\keyword{turbulent}
