% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.dist.rgh.R
\name{def.dist.rgh}
\alias{def.dist.rgh}
\title{Definition function: Aerodynamic roughness length}
\usage{
def.dist.rgh(distZaxsMeas, distObkv, veloXaxs, veloFric, RngStblObkv = c(-2,
  1))
}
\arguments{
\item{\code{distZaxsMeas}}{Measurement height and of class "numeric". [m]}

\item{\code{distObkv}}{Monin-Obukhov length and of class "numeric". [m]}

\item{\code{veloXaxs}}{Mean along-axis horizontal wind speed and of class "numeric". For eddy4R uses case, \code{veloXaxs} after pitch and roll rotation and azimuth rotation into the mean wind [m s-1]}

\item{\code{veloFric}}{Friction velocity and of class "numeric". [m s-1]}

\item{\code{RngStblObkv}}{An object of class "numeric" containing the range (minimum and maximum) of the stability parameter. Defaults to  RngStblObkv = c(-2, 1).[-]}
}
\value{
Estimated aerodynamic roughness length. [m]
}
\description{
Function defintion. Aerodynamic roughness length.
}
\examples{
def.dist.rgh (distZaxsMeas = 13.5, distObkv = 45, veloXaxs = 3, veloFric = 0.8, RngStblObkv = c(-2, 1))
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com} \cr
Natchaya Pingintha-Durden
}
\references{
Businger, J. A., Wyngaard, J. C., Izumi, Y., and Bradley, E. F.: Flux-pro<U+FB01>le relationships in the atmospheric surface layer, Journal of the Atmospheric Sciences, 28, 181<U+2013>189, 1971. \cr
Foken, T.: Micrometeorology, Springer, Berlin, Heidelberg, 2008. \cr
H<U+00F6>gstr<U+00F6>m, U.: Non-dimensional wind and temperature pro<U+FB01>les in the atmospheric surface layer: a re-evaluation, Boundary-Layer Meteorology, 42, 55<U+2013>78, 1988. \cr
}
\seealso{
Currently none
}
\keyword{eddy-covariance,}
\keyword{flux}
\keyword{length,}
\keyword{roughness}
\keyword{turbulent}

