% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.flux.chem.R
\name{def.flux.chem}
\alias{def.flux.chem}
\title{Define Chemical Species Flux}
\usage{
def.flux.chem(imfl, mn, corr, base, spcs, spcsRMM)
}
\arguments{
\item{imfl}{immediate fluctuations, created during "BASE STATE AND DEVIATIONS" step. [data.frame]}

\item{mn}{mean of input data. created during "TIME SERIES AVERAGES" step [data.frame]}

\item{corr}{correlations [data.frame]}

\item{base}{base state of data, created during "BASE STATE AND DEVIATIONS" step. [data.frame]}

\item{spcs}{names of species as molecular formulae, using IUPAC style where approprate
e.g \code{c("NO","NO2","C6H6")} [character vector]}

\item{spcsRMM}{relative molecular mass of spcs in kg mol-1. must be of equal length to spcs
e.g \code{c(0.03001,0.46005,0.07811)}. See \code{eddy4R.base::IntlNatu$Molm*} for some built in spcsRMM}
}
\value{
list containing the updated imfl, mn and cor objects
}
\description{
General Version of the CH4 flux calculation. Can use built in species RMM from IntlNatu or supply custom values
}
\author{
W. S. Drysdale
}
