% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.foot.k04.R
\name{def.foot.k04}
\alias{def.foot.k04}
\title{Definition function: Flux footprint after Kljun et a. (2004), Metzger et al. (2012)}
\usage{
def.foot.k04(
  angZaxsErth,
  distReso,
  veloYaxsHorSd,
  veloZaxsHorSd,
  veloFric,
  distZaxsMeasDisp,
  distZaxsRgh,
  distZaxsAbl = 1000,
  thsh = 0.8,
  univFunc = 0
)
}
\arguments{
\item{\code{angZaxsErth}}{wind direction to rotate the inertial footprint matrix [Deg]}

\item{\code{distReso}}{cell size of result grid [m]}

\item{\code{veloXaxsYaxs}}{horizontal wind speed [m/s]}

\item{\code{veloYaxsHorSd}}{crosswind fluctuations [m/s]}

\item{\code{veloFric}}{friction velocity [m/s]}

\item{\code{distZaxsMeasDisp}}{height of measurement - displacement [m]}

\item{\code{distObkv}}{Obukhov Length [m]}

\item{\code{distZaxsAbl}}{boundary layer height}

\item{\code{thsh}}{threshold for cumulative footprint extent}

\item{\code{univFunc}}{integral over the stability-dependent universal function to make the log-wind-profile applicable to different atmospheric stratifications; from def.func.univ()}
}
\value{
Footprint weight matrix
}
\description{
Flux footprint after Kljun et a. (2004), Metzger et al. (2012).
}
\examples{
Currently none
}
\references{
Currently none
}
\seealso{
Currently none
}
\author{
#' Stefan Metzger \email{eddy4R.info@gmail.com} \cr
Chris Florian
}
\keyword{area}
\keyword{footprint,}
\keyword{source}
