% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.func.univ.R
\name{def.func.univ}
\alias{def.func.univ}
\title{Definition function: Integral over the universal function}
\usage{
def.func.univ(distZaxsMeas, distObkv, RngStblObkv = c(-2, 1))
}
\arguments{
\item{\code{distZaxsMeas}}{Measurement height and of class "numeric". [m]}

\item{\code{distObkv}}{Monin-Obukhov length and of class "numeric". [m]}

\item{\code{RngStblObkv}}{An object of class "numeric" containing the range (minimum and maximum) of the stability parameter. Defaults to  RngStblObkv = c(-2, 1).[-]}
}
\value{
The function returns a named list containing the following:\cr
univFunc = A list object of class "numeric" containing the integral over the universal function value. \cr
flagStblObkv = A list object of class "numeric" containing the flag which indicate how the universal function being calculated (0 = calculated using the stability parameter within defined in \code{RngStblObkv}, 1 = calculated using the minimum stability parameter defined in \code{RngStblObkv}, and -1 = calculated using the maximum stability parameter defined in \code{RngStblObkv}). \cr
}
\description{
Function defintion. Calculation of the integral over the universal function after Businger et. al. (1971) in the form of Högström (1988).
}
\examples{
def.func.univ(distZaxsMeas = 13.5,distObkv = 45 ,RngStblObkv = c(-2,1))
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com} \cr
Natchaya Pingintha-Durden
}
\references{
Businger, J. A., Wyngaard, J. C., Izumi, Y., and Bradley, E. F.: Flux-proﬁle relationships in the atmospheric surface layer, Journal of the Atmospheric Sciences, 28, 181–189, 1971. \cr
Foken, T.: Micrometeorology, Springer, Berlin, Heidelberg, 2008. \cr
Högström, U.: Non-dimensional wind and temperature proﬁles in the atmospheric surface layer: a re-evaluation, Boundary-Layer Meteorology, 42, 55–78, 1988. \cr
}
\seealso{
Currently none
}
\keyword{eddy-covariance,}
\keyword{flux,}
\keyword{function}
\keyword{turbulent}
\keyword{universal}

