% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.heat.air.wet.R
\name{def.heat.air.wet}
\alias{def.heat.air.wet}
\title{Definition function: Calculation of the volumetric heat capacity}
\usage{
def.heat.air.wet(densMoleAirDry, densMoleH2o)
}
\arguments{
\item{densMoleAirDry}{A vector containing the dry air mole density, of class "numeric". [mol m-3]}

\item{densMoleH2o}{A vector containing the H2O mole density, of class "numeric". [mol m-3]}
}
\value{
The returned object is the volumetric heat capacity. [kg m-1 s-2 K-1] = specific heat [J kg-1 K-1] x mole density [mol m-3] x mole mass [kg mol-1]
}
\description{
Function definition. Calculation of the volumetric heat capacity
}
\examples{
Example 1, this will cause an error message due to densMoleAirDry, densMoleH2o have no units:
def.heat.air.wet(densMoleAirDry = 41.9054, densMoleH2o = 0.01342256)
Example 2, assign values and units to variables first, the function should run ok.
densMoleAirDry = 41.9054; attributes(densMoleAirDry)$unit <- "mol m-3"
densMoleH2o = 0.01342256; attributes(densMoleH2o)$unit <- "mol m-3"
def.heat.air.wet(densMoleAirDry, densMoleH2o)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none.
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\keyword{flux,}
\keyword{heat}
\keyword{sensible}
\keyword{temperature,}
