% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.heat.h2o.gas.temp.R
\name{def.heat.h2o.gas.temp}
\alias{def.heat.h2o.gas.temp}
\title{Definition function: Calculation of the latent heat of vaporization}
\usage{
def.heat.h2o.gas.temp(tempAir)
}
\arguments{
\item{tempAir}{A vector containing the air temperature, of class "numeric". [K]}
}
\value{
The returned object is the latent heat of vaporization. [J kg-1] == [m2 s-2]
}
\description{
Function definition. Calculation of the latent heat of vaporization.
}
\examples{
Example 1, this will cause an error message due to tempAir has no unit:
def.heat.h2o.gas.temp(tempAir = 268)
Example 2, assign values and units to variables first, the function should run ok.
tempAir <- 268
attributes(tempAir)$unit <- "K"
def.heat.h2o.gas.temp(tempAir)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
Foken 2008, Eq 2.55
}
\seealso{
Currently none.
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\keyword{gas}
\keyword{heat}
\keyword{liquid}
\keyword{phase,}
\keyword{temperature,}
