% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.natu.air.wet.R
\name{def.natu.air.wet}
\alias{def.natu.air.wet}
\title{Definition function: Calculation of the thermodynamic properties of a mix of dry air and water vapor}
\usage{
def.natu.air.wet(rtioMoleDryH2o)
}
\arguments{
\item{rtioMoleDryH2o}{A vector containing the water vapor dry mole fraction, of class "numeric". [molH2o mol-1Dry]}
}
\value{
The returned object is a data frame with the same number of observations as rtioMoleDryH2o, containing the following variables:
   $cpWet: specific heat at constant pressure of dry air + water vapor mix. [J kg-1 K-1]
   $cvWet: specific heat at constant volume of dry air + water vapor mix. [J kg-1 K-1]
   $rsWet: specific gas constant. [J kg-1 K-1]
   $gmmaWet: ratio of specific heat at constant pressure to specific heat at constant volume. [-]
   $kppaWet: Kappa exponent for ideal gas law (Poisson's equation of adiabatic change). [-]
}
\description{
Function definition. Calculation of the thermodynamic properties of a mix of dry air and water vapor.
}
\examples{
Example 1, this will cause an error message due to tempAir has no unit:
def.natu.air.wet(rtioMoleDryH2o = 0.0003)
Example 2, assign values and units to variables first, the function should run ok.
rtioMoleDryH2o <- 0.0003
attributes(rtioMoleDryH2o)$unit <- "molH2o mol-1Dry"
def.natu.air.wet(rtioMoleDryH2o)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none.
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\keyword{adiabatic}
\keyword{change,}
\keyword{constant}
\keyword{thermodynamics,}
\keyword{vapor,}
\keyword{water}
