% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.nois.R
\name{def.nois}
\alias{def.nois}
\title{Definition function: Determination of noise and detection limit for eddy-covariance turbulent fluxes}
\usage{
def.nois(
  dataTest,
  dataRefe,
  idxFlux,
  AlgBase,
  corTempPot = TRUE,
  presTempPot = NULL,
  ConfLvl = 0.95,
  CritMax = 0.01,
  PltfEc = "towr",
  flagCh4 = TRUE
)
}
\arguments{
\item{\code{dataTest}}{A dataframe containing the input data, of class "numeric". [user-defined]}

\item{\code{dataRefe}}{A vector containing the reference fluxes, of class "numeric". [user-defined]}

\item{\code{idxFlux}}{Names of columnns in \code{dataRefe} that contain the fluxes for which the detection limit calculation is performed, of class "character" (column name). [-]}

\item{\code{AlgBase}}{c("mean", "trnd", "ord03") algorithm used to determine base state, where \cr
"mean" is the simple algorithmic mean, \cr
"trnd" is the least squares linear (1st order) trend, and \cr
"ord03" is the least squares 3rd order polynomial fit.}

\item{\code{corTempPot}}{Logical value indicating whether or not to use potential temperature in the flux calculation. Defaults to TRUE. [-]}

\item{\code{presTempPot}}{A vector containing the air pressure data that will be used in the calculation when \code{corTempPot} = TRUE. Of class "numeric" or "integer" and of the same length as \code{dataTest} or single entry. [Pa]}

\item{\code{ConfLvl}}{The confidence level at which the detection limit is calculated. Of class "numeric", defaults to 0.95. [-]}

\item{\code{CritMax}}{The stop criterion for the iteration. Of class "numeric", defaults to 0.01 (i.e., 1 percent change among subsequent runs). [-]}

\item{\code{PltfEc}}{A specifier indicating which eddy covariance platform data are processed. Should be either "airc" or "towr". Defaults to "airc". [-]}

\item{\code{flagCh4}}{A logical indicating whether or not methane flux is processed. Defaults to TRUE. [-]}
}
\value{
A list containing the noise bias, noise dispersion, detection limit and signal-to-noise ratio.
}
\description{
Function defintion. Determine the noise bias, noise dispersion, detection limit and signal-to-noise ratio for turbulent fluxes using the "random shuffle" technique (Billesbach, 2011).
}
\examples{
Currently none
}
\references{
Billesbach, D.P. (2011) Estimating uncertainties in individual eddy covariance flux measurements: A comparison of methods and proposed new method. Agricultural and Forest Meteorology, 151, 394-405.
}
\seealso{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com} \ cr
Hongyan Luo \email{eddy4R.info@gmail.com}
}
\keyword{detection}
\keyword{eddy-covariance,}
\keyword{flux,}
\keyword{limit}
\keyword{turbulent}
