% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.pf.derv.coef.R
\name{def.pf.derv.coef}
\alias{def.pf.derv.coef}
\title{Definition function: Derive planar fit coefficients}
\usage{
def.pf.derv.coef(veloXaxs, veloYaxs, veloZaxs)
}
\arguments{
\item{veloXaxs}{numeric vector containing the horizontal (met: u) wind velocities in ms-1}

\item{veloYaxs}{numeric vector containing the horizontal (met: v) wind velocities in ms-1}

\item{veloZaxs}{numeric vector containing the vertical (met: z) wind velocities in ms-1}
}
\value{
A data.frame \code{coefPf} containing the planar fit coefficients (\code{angEnuYaxs} = Alpha,\code{angEnuXaxs} = Beta,\code{Ofst} = Offset).
}
\description{
Function defintion. Determines planar fit coefficients on wind vector in right-hand body coordinate system, e.g. CSAT3: Positive from front, left, below. Met: positive from west, south and below. The coefficients don't have to be in the East-North-Up (ENU) coordinate system (AngEnuXaxs); however, the names are consistent with NEON data terminology and in our data pipeline are in ENU.
}
\examples{
Currently none
}
\references{
Wilczak et al., 2001
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Ke Xu \email{xuke2012abroad@gmail.com},
Stefan Metzger \email{eddy4R.info@gmail.com}
David Durden \email{eddy4R.info@gmail.com}
}
\keyword{coordinate}
\keyword{fit,}
\keyword{planar}
\keyword{rotation,}
\keyword{transformation}
