% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.pf.rot.R
\name{def.pf.rot}
\alias{def.pf.rot}
\title{Definition function: Perform planar fit transformation}
\usage{
def.pf.rot(
  veloWind = data.frame(veloXaxs, veloYaxs, veloZaxs),
  AngEnuYaxs = 0,
  AngEnuXaxs = 0,
  Ofst = 0
)
}
\arguments{
\item{veloWind}{data.frame containing the horizontal (u,v) and vertical (z) wind velocities in ms-1}

\item{AngEnuYaxs}{The planar fit coefficient for pitch angle rotation - alpha (al)}

\item{AngEnuXaxs}{The planar fit coefficient for roll angle rotation - beta (be)}

\item{Ofst}{The planar fit coefficient for vertical velocity offset (b0)}
}
\value{
A data.frame \code{veloWindPf} containing data rotated using the planar fit transformation with the same dimensions as \code{veloWind}.
}
\description{
Function defintion. Performs planar fit transformation on wind vector in right-hand body coordinate system, e.g. CSAT3: Positive from front, left, below.
}
\examples{
Currently none
}
\references{
Wilczak et al., 2001
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Ke Xu \email{xuke2012abroad@gmail.com},
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\keyword{coordinate}
\keyword{fit,}
\keyword{planar}
\keyword{rotation,}
\keyword{transformation}
