% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.rot.ang.zaxs.erth.R
\name{def.rot.ang.zaxs.erth}
\alias{def.rot.ang.zaxs.erth}
\title{Definition function: Streamwise rotation of the wind vector}
\usage{
def.rot.ang.zaxs.erth(inp)
}
\arguments{
\item{inp}{A data frame containing the wind vector in meteorological convention with the variables veloXaxs (latitudinal wind speed, positive from west), veloYaxs (longitudinal wind speed, positive from south), and veloZaxs (vertical wind speed, positive from below) of class "numeric, each with unit attribute. [m s-1]}
}
\value{
The returned object is a list containing the elements data and rot.
data is a dataframe with the same number of observations as the function call inputs. It contains the wind vector variables in streamwise convention veloXaxsHor (streamwise wind speed, positive from front), veloYaxsHor (cross-wind speed, positive from left) and veloZaxsHor (vertical wind speed, positive from below) [m s-1], and the wind direction angZaxsErth [rad], each of class "numeric and with unit attribute.
rot is a list with the objects used in the rotation, averaged over all observations in the function call inputs. It contains the mean wind direction angZaxsErth [rad], the resulting rotation matrix mtrxRot01 [-] and the transpose of the rotation matrix mtrxRot02 [-], each of class "numeric and with unit attribute. It should be noted that angZaxsErth is calculated by first averaging each horizontal component of the wind vector, which minimizes the mean cross-wind thus satisfying conditions for footprint modeling and separating shear into stream-wise and cross-wind terms.
}
\description{
Function definition. This function rotates the wind vector into the mean wind, thus allowing to separate stream-wise and cross-wind components for subsequent use in footprint models, calculating shear stress etc. It consists of a single azimuth-rotation around the vertical axis. Any more comprehensive rotation such as double rotation or planar fit should be applied prior to calling this function.
}
\examples{
Example 1, this will cause an error message due to inp01$veloYaxs is missing:
inp01 <- base::data.frame(
  veloXaxs = c(-1.889635, -1.661724, -1.615837, -1.711132, -1.223001),
  veloZaxs = c(0.176613897, 0.184947662, 0.344331819, 0.190230311, 0.239193186)
)
attr(inp01$veloXaxs,"unit") <- "m s-1"; attr(inp01$veloZaxs,"unit") <- "m s-1"
def.rot.ang.zaxs.erth(inp = inp01)
base::rm(inp01)
Example 2, make sure to assign all variables and units, the function should run ok.
inp02 <- base::data.frame(
  veloXaxs = c(-1.889635, -1.661724, -1.615837, -1.711132, -1.223001),
  veloYaxs = c(1.365195, 1.277106, 1.394891, 1.180698, 1.283836),
  veloZaxs = c(0.176613897, 0.184947662, 0.344331819, 0.190230311, 0.239193186)
)
attr(inp02$veloXaxs,"unit") <- "m s-1"; attr(inp02$veloYaxs,"unit") <- "m s-1"; attr(inp02$veloZaxs,"unit") <- "m s-1"
out02 <- def.rot.ang.zaxs.erth(inp = inp02)
utils::str(out02)
base::rm(inp02, out02)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none.
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\keyword{footprint,}
\keyword{mean}
\keyword{rotation,}
\keyword{shear}
\keyword{stream-wise,}
\keyword{wind,}
