% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.rot.smp.R
\name{def.rot.smp}
\alias{def.rot.smp}
\title{Define Simple Rotation of Wind Vectors}
\usage{
def.rot.smp(data, MethRot = c("single", "double", "none")[2])
}
\arguments{
\item{data}{data.frame containing veloXaxs, veloYaxs and veloZaxs \cr
vectors should be defined correspondingly as u/v/w == E/N/U == x/y/z [data.frame]}

\item{MethRot}{method of rotation to be used, one of: \itemize{
               \item "single" - rotate into the mean wind
               \item "double" - apply single rotation and additionally rotate to align w and z (minimise w)
               \item "none" - perform no rotation
}}
}
\value{
Data object with rotated wind vectors names as veloXaxs, veloYaxs and veloZaxs
}
\description{
Perform various types of wind vector rotations incl. single, double
}
\references{
Code adapted from REYNFlux_P5 (35ceda9) and flow.turb.tow.neon.dp04.r (914a9e9) \cr
 Stefan Metzger / Dave Durden / Natchaya P-Durden / Cove Sturtevant / Ke Xu
}
\author{
David Durden \email{ddurden@battelleecology.org}

W. S. Drysdale
}
