% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.spcs.name.R
\name{def.spcs.name}
\alias{def.spcs.name}
\title{Name Species}
\usage{
def.spcs.name(spcs, type = c("mole", "mass", "kin")[1])
}
\arguments{
\item{spcs}{species name such that FD_mole_species corresponds to a column header in data [character]}

\item{type}{one of c("mole","mass","kin") for each type of name required}
}
\description{
generate eddy4r style names for chemical species
}
\author{
W. S. Drysdale
}
