% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.spec.cosp.R
\name{def.spec.cosp}
\alias{def.spec.cosp}
\title{Definition function: Generate cospectra}
\usage{
def.spec.cosp(dataSpec, dataTimeDomn, CoefConvEngy = 1)
}
\arguments{
\item{dataSpec}{data.frame or matrix frequency space data, raw spectra as complex numbers}

\item{dataTimeDomn}{data.frame of time domain data, identical variables and length as dataSpec}

\item{CoefConvEngy}{A single factor (specific heat at constant pressure * air density) for conversion from kinematic to energetic units; only used for comparison with Reynolds-Covariance, not for output; defaults to 1}
}
\value{
A list \code{rpt} that contains the cospectral data (\code{rpt$dataCosp}) and the summed cospectra (\code{rpt$sumCosp}), which includes evaluation against total covariance and the relative difference
}
\description{
Generate cospectra.
}
\examples{
Currently none
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}

David Durden
}
\keyword{FFT,}
\keyword{Fast}
\keyword{Fourier}
\keyword{Transform,}
\keyword{cospectra}
\keyword{spectral,}
