% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.spec.freq.cnst.R
\name{def.spec.freq.cnst}
\alias{def.spec.freq.cnst}
\title{Definition function: 63\% frequency constant after Aubinet (2012) Eq. 4.22}
\usage{
def.spec.freq.cnst(inpTime)
}
\arguments{
\item{inpPrd}{numeric, time period in seconds}
}
\value{
A numeric vector \code{freqOut} of equivalent frequencies to the input time constant
}
\description{
63\% frequency constant after Aubinet (2012) Eq. 4.22.
}
\examples{

inpTime <- seq(1, 3600, by = 10)
def.spec.freq.cnst(inpTime = inpTime)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
Foken T., Leuning R., Oncley S.R., Mauder M., Aubinet M. (2012) Corrections and Data Quality Control. In: Aubinet M., Vesala T., Papale D. (eds) Eddy Covariance. Springer Atmospheric Sciences. Springer, Dordrecht. https://doi.org/10.1007/978-94-007-2351-1_4
Foken, Thomas. (2017). Micrometeorology. 10.1007/978-3-642-25440-6.
}
\seealso{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}

David Durden
}
\keyword{FFT,}
\keyword{Fast}
\keyword{Fourier}
\keyword{Transform,}
\keyword{spectral}
