% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.spec.freq.cut.emp.R
\name{def.spec.freq.cut.emp}
\alias{def.spec.freq.cut.emp}
\title{Definition function: Determine cutoff frequency empirically}
\usage{
def.spec.freq.cut.emp(FreqCut, idep, depe, CoefCorRefe)
}
\arguments{
\item{FreqCut}{cutoff frequency}

\item{idep}{independent variable, preferabley f, but n is possible}

\item{depe}{dependent variable, spectra or cospectra}

\item{CoefCorRefe}{reference correction factor}
}
\value{
Optimality criterion from the difference of the reference correction and the derived correction.
}
\description{
Determine cutoff frequency empirically .
}
\examples{
Currently none
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
Foken T., Leuning R., Oncley S.R., Mauder M., Aubinet M. (2012) Corrections and Data Quality Control. In: Aubinet M., Vesala T., Papale D. (eds) Eddy Covariance. Springer Atmospheric Sciences. Springer, Dordrecht. https://doi.org/10.1007/978-94-007-2351-1_4
Eugster, W. and W. Senn (1995). "A COSPECTRAL CORRECTION MODEL FOR MEASUREMENT OF TURBULENT NO2 FLUX." Boundary-Layer Meteorology 74(4): 321-340.
Foken, Thomas. (2017). Micrometeorology. 10.1007/978-3-642-25440-6.
}
\seealso{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}

David Durden
}
\keyword{FFT,}
\keyword{Fast}
\keyword{Fourier}
\keyword{Transform,}
\keyword{cut-off}
\keyword{frequency,}
\keyword{function}
\keyword{spectral,}
\keyword{transfer}
