% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.spec.high.freq.cor.R
\name{def.spec.high.freq.cor}
\alias{def.spec.high.freq.cor}
\title{Definition function: function to determine the temporally resolved covariance from discrete wavelet transform and frequency response correction}
\usage{
def.spec.high.freq.cor(
  sclrDwt,
  veloZaxsDwt,
  sclrSpecPowr,
  FreqSamp = veloZaxsDwt@attr.X$tsp[3],
  freq = 2^(seq(veloZaxsDwt@level - 1, 0)) * FreqSamp/2^veloZaxsDwt@level,
  freqPeak = rep(10, 2),
  idxFreqPeak = rep(ceiling(length(freq)/2), 2),
  sclrCnst = 1,
  idxData = seq(1, length(sclrDwt@series)),
  qfWave
)
}
\arguments{
\item{sclrDwt}{wavelets object output, discrete wavelet transform output object for a given scalar/variable (e.g., rtioMoleDryH2o)}

\item{veloZaxsDwt}{wavelets object output, discrete wavelet transform output object for a vertical wind speed (i.e., w')}

\item{sclrSpecPowr}{vector of spectral power estimates calculated for sclrDwt, derived from wavelet coefficients from dwt.R output following Eq. 7 in NK12}

\item{FreqSamp}{Sampling frequency, defaults to 20 Hz}

\item{freq}{vector of frequencies for wavelet spectral results}

\item{freqPeak}{vector (length = 2) for peak frequency values of scalar variable and vertical wind speed variable, respectively}

\item{idxFreqPeak}{vector (length = 2) for index from 'freq' vector for peak frequency values of both scalar variable and vertical wind speed variable, respectively}

\item{sclrCnst}{scaling constant in scientific notation (e.g., 1e06, 1e-04), needed to account for rounding in idwt.R in 'wavelets' package}

\item{idxData}{index of time series data to use and modify. Only needed if zero-padding is being done. Currently just defaults to every observation in dataset.}

\item{qfWave}{Wavelet flag: process (0) or not (1) depending on if number of data points missing >10%}
}
\value{
A list containing output from NK12 high frequency correction, including a correction coefficients for use in flux data (coefCor)
}
\description{
Definition function. Function to determine the temporally resolved variance/covariance from discrete wavelet transform and frequency response correction described in Nordbo and Katul (2012)
}
\examples{
Currently none.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none.
}
\author{
Adam Young \email{younga1@battelleecology.org}
David Durden
Stefan Metzger
}
\keyword{NK12,}
\keyword{Wavelet,}
\keyword{correction}
\keyword{cospectrum,}
\keyword{frequency}
\keyword{response}
\keyword{spectrum,}
