% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.spec.name.R
\name{def.spec.name}
\alias{def.spec.name}
\title{Name Species}
\usage{
def.spec.name(species, type = c("mole", "mass", "kin")[1])
}
\arguments{
\item{type}{}
}
\description{
generate eddy4r style names for chemical species
}
\author{
W. S. Drysdale
}
