% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.spec.peak.modl.R
\name{def.spec.peak.modl}
\alias{def.spec.peak.modl}
\title{Definition function: Function to define nonlinear model to fit to veloZaxsHor spectra to determine frequency peak}
\usage{
def.spec.peak.modl(para, freq)
}
\arguments{
\item{para}{Parameters to be fit model via optimization.}

\item{freq}{frequency vector as independent variable.}
}
\value{
Vector of spectra values for set of parameters.
}
\description{
# Simple non-linear model used in NK12 to model weighted vertical wind spectra and then use modeled spectral curve to identify frequency that spectral peak occurs at. Based on unpublished Matlab code provide to Stefan Metzger from authors of NK12.
}
\examples{
Currently none.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none.
}
\author{
Adam Young \email{younga1@battelleecology.org}
}
\keyword{Currently}
\keyword{none.}
