% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.spec.peak.ogiv.R
\name{def.spec.peak.ogiv}
\alias{def.spec.peak.ogiv}
\title{Definition function: Determine spectral peak using an Ogive method}
\usage{
def.spec.peak.ogiv(
  FreqPeak,
  idep,
  depe,
  MethSpec = c("spec", "cosp")[2],
  paraStbl,
  MethWght = TRUE,
  ThshFreqRng = c(0.01, 1),
  CoefCumScal = 0.6,
  FilePlot = NULL
)
}
\arguments{
\item{FreqPeak}{frequency f at which fCO(f) reaches its maximum value}

\item{idep}{independent variable, preferabley f, but n is possible}

\item{depe}{dependent variable, spectra or cospectra}

\item{MethSpec}{spectrum or cospectrum?}

\item{paraStbl}{stability parameter}

\item{MethWght}{use frequency-weighted (co)spectrum?}

\item{ThshFreqRng}{frequency range for determining optimiality criterion}

\item{CoefCumScal}{cumulative flux contribution for which measured (co)-spectrum is scaled to model (co)-spectrum}

\item{FilePlot}{generate plot?}
}
\value{
The optimality criterion and the cumulative (Ogive) correction coefficient is returned for the calculated independent variable (frequency/wavenumber) scaling index for the dependent variables (spectra or cospectra). If \code{FilePlot} is provided, a plot will be generated and output.
}
\description{
Determine spectral peak using an Ogive method.
}
\examples{
Currently none
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}

David Durden
}
\keyword{FFT,}
\keyword{Fast}
\keyword{Fourier}
\keyword{Ogive}
\keyword{Transform,}
\keyword{spectral,}
