% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.spec.peak.R
\name{def.spec.peak}
\alias{def.spec.peak}
\title{Definition function: Function to determine frequency that peak normalized spectral values occur for a given variable.}
\usage{
def.spec.peak(specPowrWght, para, freq)
}
\arguments{
\item{specPowrWght}{Spectra to fit model to}

\item{para}{Parameters to be fit model via optimization.}

\item{freq}{frequency vector as independent variable.}
}
\value{
List of spectra peak freq values and indices, as well as best fit parameters and interpolated results.
}
\description{
Function to define nonlinear model to fit to veloZaxsHor spectra and determine frequency peak. The output from this model is then used with 'optim' to find the frequency at which vertical wind speak spectral power reaches it's peak value.
}
\examples{
Currently none.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none.
}
\author{
Adam Young \email{younga1@battelleecology.org}
}
\keyword{Currently}
\keyword{none.}
