% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.spec.plot.cosp.R
\name{def.spec.plot.cosp}
\alias{def.spec.plot.cosp}
\title{Definition function: Plot cospectra}
\usage{
def.spec.plot.cosp(
  FileOut,
  idep,
  depe,
  DscrPlot,
  Labx = NA,
  Laby = NA,
  Colr = c(1, 2, 1),
  Limx = NULL,
  Limy = NULL,
  MethWght = TRUE,
  MethDistWave = FALSE,
  CoefNorm = NULL,
  CoefScalPlot = 1,
  MethModlSpec = TRUE,
  paraStbl,
  FreqPeakCosp = NULL,
  FreqRngEval = NULL,
  MethPlot = TRUE
)
}
\arguments{
\item{FileOut}{(no default) output location and filename for plot}

\item{idep}{(no default) one independent variable, e.g. frequencey, wavenumber...}

\item{depe}{(no default) spectra of up to three dependent variables, same length as idep}

\item{DscrPlot}{(no default) one description (character or expression) for each variable for the title}

\item{Labx}{(NA) description for abscissa}

\item{Laby}{(NA) description for ordinate}

\item{Colr}{(c(1,2,1)) colors for plotting of measurement, model and evaluation rang}

\item{Limx}{(NULL) limits for abscissa}

\item{Limy}{(NULL) limits for ordinate}

\item{MethWght}{(T) weight the spectra for independent variable (wavelength / frequency)?}

\item{MethDistWave}{(F) is wavelength used as independent variable?}

\item{CoefNorm}{(NULL) normalizes sum(cospectrum) to a given quantity, e.g. unity or the total flux per variable, else NULL}

\item{CoefScalPlot}{(1) allows to plot measured cospectrum with scaling factor that is different from norm}

\item{MethModlSpec}{(T) shall model cospetra be calculated and plotted?}

\item{paraStbl}{(no default) atmospheric stability to choose the reference cospectrum}

\item{FreqPeakCosp}{(NULL) frequency at which fCO(f) reaches its maximum value; will be determined from measured data if NULL}

\item{FreqRngEval}{(NULL) range to assess sum((cospectrum - model)[setFreq]) / sum(model) [%]; only valid for equidistant data}

\item{MethPlot}{logical to determine if plot should be output to FileOut}
}
\value{
The Correction Coefficient (\code{CoefCor}) is returned if \code{FreqRngEval} is defined. Additionally, if \code{MethPlot == TRUE} plots will be generated and output to \code{FileOut}.
}
\description{
Plot cospectra.
}
\examples{
Currently none
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
Kaimal (1972)
Massman, 2005 (in Lee, 2005)
}
\seealso{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}

David Durden
}
\keyword{FFT,}
\keyword{Fast}
\keyword{Fourier}
\keyword{Transform,}
\keyword{cospectra}
\keyword{spectral,}
