% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.spec.plot.ogiv.R
\name{def.spec.plot.ogiv}
\alias{def.spec.plot.ogiv}
\title{Definition function: Plot ogives}
\usage{
def.spec.plot.ogiv(
  FilePlot,
  idepRaw,
  cospRaw,
  idepSmth,
  cospSmth,
  ThshFluxLvl,
  PrdFluxLvl,
  DscrPlot = NULL,
  Labx = NULL,
  Laby = NULL,
  Colr = c("grey", 1, 1),
  Limx = NULL,
  Limy = NULL,
  MethWght = F,
  MethDistWave = TRUE,
  CoefNorm = 1
)
}
\arguments{
\item{FilePlot}{output location and filename for plot (no default)}

\item{idepRaw}{one independent variable for the raw cospectra, e.g. frequencey, wavenumber...(no default)}

\item{cospRaw}{raw cospectra of up to four dependent variables, same length as idepRaw (no default)}

\item{idepSmth}{one independent variable for the smoothed cospectra, e.g. frequencey, wavenumber (no default)}

\item{cospSmth}{smoothed cospectra of up to four dependent variables, same length as idepRaw (no default)}

\item{ThshFluxLvl}{Flux contribution level to evaluate e.g. corresponding wavelength (no default)}

\item{PrdFluxLvl}{flux averaging period to evaluate corresponding contribution level (no default)}

\item{DscrPlot}{one description (character or expression) for each variable for the title (no default)}

\item{Labx}{description for abscissa (NA)}

\item{Laby}{description for ordinate (NA)}

\item{Colr}{colors for plotting of measurement, model and evaluation range (c(1,2,1))}

\item{Limx}{limits for abscissa (NULL)}

\item{Limy}{limits for ordinate (NULL)}

\item{MethWght}{weight the cospectra for independent variable (wavelength / frequency)? (FALSE)}

\item{MethDistWave}{is wavelength used as independent variable? (TRUE)}

\item{CoefNorm}{normalizes max(cospectrum) to a given quantity, e.g. total flux per variable, else unity (Default = 1)}
}
\value{
Outputs Ogive plots to \code{FilePlot} and returns a vector or data.frame containing \code{DistWaveFluxLvl} wavelength corresponding to given contribution level and \code{FluxLvlPrd} determine flux contribution level for given wavelength for each cospectra.
}
\description{
Plot ogives. Note: Accumulation only valid with raw data or equidistant binned data!
}
\examples{
Currently none
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}

David Durden
}
\keyword{FFT,}
\keyword{Fast}
\keyword{Fourier}
\keyword{Ogive}
\keyword{Transform,}
\keyword{spectral,}
