% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.spec.plot.powr.R
\name{def.spec.plot.powr}
\alias{def.spec.plot.powr}
\title{Definition function: Plot power spectra of up to three variables}
\usage{
def.spec.plot.powr(
  FileOut,
  idep,
  depe,
  DscrVar,
  Labx = NA,
  Laby = NA,
  Colr = NULL,
  Limx = NULL,
  Limy = NULL,
  Ofst = NULL,
  CoefPowrSlp = NULL,
  FreqSub = c(0.05, Inf)
)
}
\arguments{
\item{FileOut}{character, output location and filename for plot}

\item{idep}{one independent variable, e.g. frequencey, wavenumber...}

\item{depe}{spectra of up to three dependent variables, same length as idep}

\item{DscrVar}{one description (character or expression) for each variable for the legend}

\item{Labx}{character (NA), description for abscissa}

\item{Laby}{character (NA), description for ordinate}

\item{Colr}{(NULL) colors for plotting, as many colors as dependent variables}

\item{Limx}{(NULL) limits for abscissa}

\item{Limy}{(NULL) limits for ordinate}

\item{Ofst}{(NULL) offsets different variables from each other to improve legibility (only if CoefPowrSlp is set)}

\item{CoefPowrSlp}{(NULL) coefficient of the power law c("-5/3", "-2/3") for unweighted / weighted spectra}

\item{FreqSub}{frequency range of initial subrange, used to fit the power law which defaults to (0.05,Inf)}
}
\value{
Plot output to FileOut
}
\description{
Plot power spectra of up to three variables.
}
\examples{
Currently none
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}

David Durden
}
\keyword{FFT,}
\keyword{Fast}
\keyword{Fourier}
\keyword{Transform,}
\keyword{spectral}
