% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.spec.tfun.sigm.R
\name{def.spec.tfun.sigm}
\alias{def.spec.tfun.sigm}
\title{Definition function: Sigmoidal transfer function (Lorentzian)}
\usage{
def.spec.tfun.sigm(FreqCut, Freq)
}
\arguments{
\item{FreqCut}{half-power / cut-off frequency}

\item{Freq}{frequency}
}
\value{
A Sigmoidal transfer function for the defined frequencies and cut-off frequency.
}
\description{
Sigmoidal transfer function (Lorentzian) after Eugster and Senn (1995) in Aubinet et al. (2012) Eq. 4.21.
}
\examples{
Currently none
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
Foken T., Leuning R., Oncley S.R., Mauder M., Aubinet M. (2012) Corrections and Data Quality Control. In: Aubinet M., Vesala T., Papale D. (eds) Eddy Covariance. Springer Atmospheric Sciences. Springer, Dordrecht. https://doi.org/10.1007/978-94-007-2351-1_4
Eugster, W. and W. Senn (1995). "A COSPECTRAL CORRECTION MODEL FOR MEASUREMENT OF TURBULENT NO2 FLUX." Boundary-Layer Meteorology 74(4): 321-340.
Foken, Thomas. (2017). Micrometeorology. 10.1007/978-3-642-25440-6.
}
\seealso{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}

David Durden
}
\keyword{FFT,}
\keyword{Fast}
\keyword{Fourier}
\keyword{Transform,}
\keyword{spectral}
