% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.spec.time.cnst.R
\name{def.spec.time.cnst}
\alias{def.spec.time.cnst}
\title{Definition function: 63\% time constant after Aubinet (2012) Eq. 4.22}
\usage{
def.spec.time.cnst(inpFreq)
}
\arguments{
\item{inpFreq}{numeric, frequency in Hertz (Hz)}
}
\value{
A numeric vector \code{prdOut} of equivalent time periods to the input frequencies in seconds (sec)
}
\description{
63\% time constant after Aubinet (2012) Eq. 4.22.
}
\examples{

inpFreq <- seq(1, 0.02, by = -0.02)
def.spec.time.cnst(inpFreq = inpFreq)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
Foken T., Leuning R., Oncley S.R., Mauder M., Aubinet M. (2012) Corrections and Data Quality Control. In: Aubinet M., Vesala T., Papale D. (eds) Eddy Covariance. Springer Atmospheric Sciences. Springer, Dordrecht. https://doi.org/10.1007/978-94-007-2351-1_4
Foken, Thomas. (2017). Micrometeorology. 10.1007/978-3-642-25440-6.
}
\seealso{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}

David Durden
}
\keyword{FFT,}
\keyword{Fast}
\keyword{Fourier}
\keyword{Transform,}
\keyword{spectral}
