% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.stat.sta.diff.R
\name{def.stat.sta.diff}
\alias{def.stat.sta.diff}
\title{Definition function: Summary statistics, base states, instantaneous differences}
\usage{
def.stat.sta.diff(inp, refe = NULL, AlgBase = c("mean", "trnd", "ord03")[1])
}
\arguments{
\item{inp}{A data frame containing the variables for which to perform the calculations, class "numeric", each with unit attribute.}

\item{refe}{A list of reference quantities that can be supplied to overwrite internal calculations, class "numeric", each with unit attribute. Currently implemented only for refe$mean in combination with inp variables in units [rad].}

\item{AlgBase}{A vector of length 1 that defines the base state with respect to which instantaneous differences and standard deviations are calculated, of class "character" and no unit attribute. Contains one of AlgBase <- c("mean", "trnd", "ord03")[1] and defaults to "mean", with the additional options detrending "trnd" and 3rd-order polynomial "ord03". See ?eddy4R.base::def.base.ec() for additional details. When AlgBase is set to "trnd" or "ord03", the variable inp$idep is required, which provides the independent variable for interpolation.}
}
\value{
The returned object is a list containing the element dataframes min, max, mean, base, diff, and sd, each of class "numeric" and with unit attributes. The elements min, max and mean are the minimum, maximum and mean of the variables in inp with a single observation. The element base is the base state for each of the variables in inp, with a single observation (AlgBase == "mean") or the same number of observations as inp (AlgBase %in% c("trnd", "ord03")). The element diff are the point-by-point differences from the selected base state for each of the variables in inp, with the same number of observations as inp. The element sd are the standard deviations of diff for each of the variables in inp, with a single observation. It should be noted that the mean (and base state for AlgBase == "mean") for angular quantities with unit [rad] is computed from 1) point-wise unit vector decomposition to polar coordinates [-], 2) averaging the polar coordinates [-], and 3) re-composing the mean polar coordinates to Cartesian angle [rad]. In the special case of wind direction (example: angZaxsErth) it is recommended to supply the argument refe$mean$angZaxsErth to the function call as provided from ?def.rot.ang.zaxs.erth. This ensures consistency with downstream applications for footprint modeling, separating shear into stream-wise and cross-wind terms etc. (minimizes the mean cross-wind from directly averaging the wind vector horizontal components instead of unit vector components).
}
\description{
Function definition. The function calculates summary statistics (min, max, mean), selected base state (mean, trend, 3rd order polynomial), and the instantaneous differences and standard deviations with respect to the selected base state. This enables subsequent calculations (eddy-covariance turbulent fluxes, footprint modeling etc.) with respect to the selected base state.
}
\examples{
Example 1, this will cause an error message due to missing unit attribute for inp01$veloYaxs:
inp01 <- base::data.frame(
  veloXaxs = c(-1.889635, -1.661724, -1.615837, -1.711132, -1.223001),
  veloYaxs = c(1.365195, 1.277106, 1.394891, 1.180698, 1.283836),
  veloZaxs = c(0.176613897, 0.184947662, 0.344331819, 0.190230311, 0.239193186)
)
attr(inp01$veloXaxs,"unit") <- "m s-1"; attr(inp01$veloZaxs,"unit") <- "m s-1"
def.stat.sta.diff(inp = inp01)
base::rm(inp01)
Example 2, make sure to assign all variables and units, the function should run ok.
inp02 <- base::data.frame(
  veloXaxs = c(-1.889635, -1.661724, -1.615837, -1.711132, -1.223001),
  veloYaxs = c(1.365195, 1.277106, 1.394891, 1.180698, 1.283836),
  veloZaxs = c(0.176613897, 0.184947662, 0.344331819, 0.190230311, 0.239193186)
)
attr(inp02$veloXaxs,"unit") <- "m s-1"; attr(inp02$veloYaxs,"unit") <- "m s-1"; attr(inp02$veloZaxs,"unit") <- "m s-1"
out02 <- def.stat.sta.diff(inp = inp02)
utils::str(out02)
base::rm(out02)
Example 3, difference when computing angular average without and with reference (wind direction).
out03 <- def.stat.sta.diff(inp = def.rot.ang.zaxs.erth(inp = inp02)$data)
out04 <- def.stat.sta.diff(inp = def.rot.ang.zaxs.erth(inp = inp02)$data,
                           refe = base::list("mean" = base::list("angZaxsErth" = 
                                                                   def.rot.ang.zaxs.erth(inp = inp02)$rot$angZaxsErth))
                           )
utils::str(out03)
utils::str(out04)
out03$mean$angZaxsErth - out04$mean$angZaxsErth
base::rm(inp02, out03, out04)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none.
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\keyword{deviation,}
\keyword{maximum,}
\keyword{mean,}
\keyword{minimum,}
\keyword{standard}
\keyword{statistics,}
\keyword{turbulence}
