% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.temp.virt.temp.air.pres.h2o.pres.atm.R
\name{def.temp.virt.temp.air.pres.h2o.pres.atm}
\alias{def.temp.virt.temp.air.pres.h2o.pres.atm}
\title{Definition function: Calculation of the virtual temperature}
\usage{
def.temp.virt.temp.air.pres.h2o.pres.atm(tempAir, presH2o, presAtm)
}
\arguments{
\item{tempAir}{A vector containing the air temperature, of class "numeric". [K]}

\item{presH2o}{A vector containing the water vapor partial pressure, of class "numeric". [Pa]}

\item{presAtm}{A vector containing the atmospheric pressure, of class "numeric". [Pa]}
}
\value{
The returned object is the the virtual temperature. [K]
}
\description{
Function definition. Calculation of the virtual temperature.
The virtual temperature is the temperature of a moist air parcel at which a theoretical 
dry air parcel would have a total pressure and density equal to the moist parcel of air.
}
\examples{
Example 1, this will cause an error message due to tempAir, presH2o, presAtm have no units:
def.temp.virt.temp.air.pres.h2o.pres.atm(tempAir = 268, presH2o = 30, presAtm = 93344)
Example 2, assign values and units to variables first, the function should run ok.
tempAir <- 268
presH2o <- 30
presAtm <- 93344
attributes(tempAir)$unit <- "K"
attributes(presH2o)$unit <- "Pa"
attributes(presAtm)$unit <- "Pa"
def.temp.virt.temp.air.pres.h2o.pres.atm(tempAir, presH2o, presAtm)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
http://en.wikipedia.org/wiki/Virtual_temperature
}
\seealso{
Currently none.
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\keyword{pressure,}
\keyword{temperature,}
\keyword{virtual}
