% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.vari.wave.R
\name{def.vari.wave}
\alias{def.vari.wave}
\title{Definition function: funtion to determine the temporally resolved variance/covariance from continuous wavelet transform}
\usage{
def.vari.wave(spec1 = mycwt[["w_met"]]@spectrum, spec2 = NULL,
  scal = mycwt[[vari]]@scale, peri = mycwt[[vari]]@period, freq_0 = NA,
  whr_peri = NULL, fac_norm = rpt$coefNorm, flag, SI, SC)
}
\arguments{
\item{spec1}{Waves package output object spectrum, continuous wavelet transform output object complex spectrum for the first variable (typically w')denoted as \code{object1@spectrum}.}

\item{spec2}{Waves package output object spectrum, continuous wavelet transform output object complex spectrum for the second variable for cospectra denoted as \code{object2@spectrum}.}

\item{scal}{Waves package output object scale, width of the wavelet at each scale [s] denoted as \code{object1@scale}}

\item{peri}{Waves package output object period, approximate corresponding Fourier period [s] denoted as \code{object1@period}}

\item{freq_0}{vector, half-power frequencies for individual variables [Hz] for determining transfer function to correct frequency response}

\item{whr_peri}{numeric, which wavelengths/spatial scales to consider if you want to only consider high frequency values}

\item{fac_norm}{numeric, normalization factor specific to the choice of Wavelet parameters.}

\item{dfInp}{data.frame, consisting of the input data to perform the wavelet transformation}
}
\value{
A vector constaining temporally resolved variance/covariance from the continuous wavelet transform.
}
\description{
Wrapper function. funtion to determine the temporally resolved variance/covariance from continuous wavelet transform including high-frequency spectral correction and selectable low-frequency cutoff The frequency response correction using Wavelet techniques described in Norbo and Katul, 2012 (NK12)
}
\examples{
Currently none.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none.
}
\author{
David Durden \email{ddurden@battelleecology.org}
Stefan Metzger
}
\keyword{NK12,}
\keyword{Wavelet,}
\keyword{correction}
\keyword{cospectrum,}
\keyword{frequency}
\keyword{response}
\keyword{spectrum,}
