% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.wave.lorenz.calc.R
\name{def.wave.lorenz.calc}
\alias{def.wave.lorenz.calc}
\title{Definition function: Function to organize and define wavelet coefficients into a Lorenz curve. Used to find most energetic coefficients for adjustment in spectral corrections.}
\usage{
def.wave.lorenz.calc(waveCoef)
}
\arguments{
\item{waveCoef}{list of wavelet coefficients for each scale.}
}
\value{
results from Lorenz curve calculations.
}
\description{
Function to organize and define wavelet coefficients into a Lorenz curve. Used to find most energetic coefficients for adjustment in spectral corrections.
}
\examples{
Currently none.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none.
}
\author{
Adam Young \email{younga1@battelleecology.org}
}
\keyword{NK12,}
\keyword{Wavelet,}
\keyword{correction}
\keyword{cospectrum,}
\keyword{frequency}
\keyword{response}
\keyword{spectrum,}
