% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.wind.mbb.R
\name{def.wind.mbb}
\alias{def.wind.mbb}
\title{Definition function: Optimal Window Length for Moving Block Bootstrap (MBB)}
\usage{
def.wind.mbb(data)
}
\arguments{
\item{\code{scalEddy}}{A vector containing distances or times and of class "numeric". [m] or [s]}

\item{\code{data}}{A vector containing the input data. Of class "numeric" or "integer". [user-defined]}

\item{\code{veloXaxs}}{Instantaneous wind speed. Only supplied when users define scalEddy in terms of time. If provided, of class "numeric", otherwise NULL. [ms-1]}
}
\value{
Optimal Length for the Moving Block Bootstrap.
}
\description{
Function defintion. Window Length for MBB
}
\examples{
Currently none
}
\references{
S. Salesky, M. Chamecki, N. Dias (2012) Estimating random error in eddy covariance fluxes and other turbulence statistics: the filtering method .  Boundary-Layer Meteorology., 144, 113-135. \cr
}
\seealso{
Currently none
}
\author{
Kenny Pratt \email{eddy4R.info@gmail.com}
}
\keyword{Block}
\keyword{Bootstrap,}
\keyword{Length}
\keyword{Moving}
\keyword{Optimal}
\keyword{Window}
