% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.r
\name{find_F0}
\alias{find_F0}
\title{Definition function: Determine cutoff frequency empirically}
\usage{
find_F0(
  f0,
  ide = SPEout$fr_obs[SPEout$fr_whr][which(SPEout$fr_obs[SPEout$fr_whr] >= 0.01)],
  dep = SPEout$FSunfold[SPEout$fr_whr, fpo][which(SPEout$fr_obs[SPEout$fr_whr] >=
    0.01)],
  corfac_ref = corfac_out
)
}
\arguments{
\item{Currently}{none}
}
\value{
Currently none
}
\description{
Determine cutoff frequency empirically .
}
\examples{
Currently none
}
\references{
Currently none
}
\seealso{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\keyword{Currently}
\keyword{none}
