% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.r
\name{find_FX_og}
\alias{find_FX_og}
\title{Determine spectral peak using an Ogive method}
\usage{
find_FX_og(FX, IDE = SPEout$fr_obs[SPEout$fr_whr],
  DEP = SPEout$FScosp[SPEout$fr_whr, 4], SC = c("spe", "cos")[2],
  SI = OUT$REYN$mn$sigma[FILE], WEIGHT = TRUE, WHR_CRIT = c(1e-02, 1),
  crit_cum = 6e-01, plot_path = NULL, meth = c("peak", "corfac")[1])
}
\arguments{
\item{Currently}{none}
}
\value{
Currently none
}
\description{
Determine spectral peak using an Ogive method.
}
\examples{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\references{
Currently none
}
\seealso{
Currently none
}
\keyword{Currently}
\keyword{none}

