% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.r
\name{find_TF}
\alias{find_TF}
\title{Definition function: Determine half-power / cut-off frequency}
\usage{
find_TF(F0 = 5e-01, FX = fx_out$par * mean(SPEout$fr_obs/SPEout$fr_nor,
  na.rm = TRUE),
  IDE = SPEout$fr_obs[SPEout$fr_whr][which(SPEout$fr_obs[SPEout$fr_whr] >=
  1e-02)], DEP = SPEout$FScosp[SPEout$fr_whr,
  fpo][which(SPEout$fr_obs[SPEout$fr_whr] >= 1e-02)], SC = c("spe", "cos")[2],
  SI = OUT$REYN$mn$sigma[FILE], WEIGHT = FALSE, WHR_CRIT = c(1e-02, 10),
  corfac = corfac_out, BINS = 1e+03, plot_path = NULL)
}
\arguments{
\item{Currently}{none}
}
\value{
Currently none
}
\description{
Determine half-power / cut-off frequency.
}
\examples{
Currently none
}
\references{
Currently none
}
\seealso{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\keyword{Currently}
\keyword{none}
