% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprints.r
\name{footK04}
\alias{footK04}
\title{Definition function: Flux footprint after Kljun et a. (2004), Metzger et al. (2012)}
\usage{
footK04(
  angle,
  Csize,
  sigV,
  sigW,
  ustar,
  zmeas,
  z0,
  h = 1000,
  thsh = 0.8,
  Psi = 0
)
}
\arguments{
\item{Currently}{none}
}
\value{
Footprint weight matrix
}
\description{
Flux footprint after Kljun et a. (2004), Metzger et al. (2012).
}
\examples{
Currently none
}
\references{
Currently none
}
\seealso{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\keyword{area}
\keyword{footprint,}
\keyword{source}
