% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprints.r
\name{footKM01}
\alias{footKM01}
\title{Definition function: Flux footprint after Kormann & Meixner (2001)}
\usage{
footKM01(angle, Csize, hor, sigV, ustar, zmeas, Lvirt, cutoff = 1e-02)
}
\arguments{
\item{\code{angle}}{wind direction to rotate the inertial footprint matrix [Deg]}

\item{\code{Csize}}{cell size of result grid [m]}

\item{\code{hor}}{horizontal wind speed [m/s]}

\item{\code{sigV}}{crosswind fluctuations [m/s]}

\item{\code{ustar}}{friction velocity [m/s]}

\item{\code{zmeas}}{height of measurement - displacement [m]}

\item{\code{Lvirt}}{Obukhov Length [m]}

\item{\code{cutoff}}{cellwise contribution relative to peak at which footprint will be cut off}
}
\value{
Footprint weight matrix
}
\description{
Function for calclulating one flux footprint matrix at a time for the crosswind integrated footprint Kormann & Meixner, 2001 (KM01). For the crosswind distribution, a Gaussian function is used(Pasquill, 1974; Kormann and Meixner, 2001). The matrix extent is precalculated, and cut off where the footprint function reaches 1 % of its peak (Neftel, 2008).
}
\examples{
Currently none
}
\references{
#'Kormann, R., and F. X. Meixner, 2001: An analytical footprint model for non-neutral stratification. Boundary-Layer Meteorol., 99, 207-224, doi:10.1023/A:1018991015119. \cr
Pasquill, F., Atmospheric diffusion; the dispersion of windborne material from industrial and other sources. 2d ed. Chichester [Eng.] E. Horwood; New York, Halsted Press [1974], https://search.library.wisc.edu/catalog/999623699702121. \cr
Neftel, A., C. Spirig, and C. Ammann, 2008: Application and test of a simple tool for operational footprint evaluations. Environ. Pollut., 152, 644-652, doi:10.1016/j.envpol.2007.06.062. \cr
}
\seealso{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\keyword{area}
\keyword{footprint,}
\keyword{source}
