% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.r
\name{fun_TSIG}
\alias{fun_TSIG}
\title{Sigmoidal transfer function (Lorentzian)}
\usage{
fun_TSIG(freq_0, freq)
}
\arguments{
\item{Currently}{none}
}
\value{
Currently none
}
\description{
Sigmoidal transfer function (Lorentzian) after Eugster and Senn (1995) in Aubinet et al. (2012) Eq. 4.21.
}
\examples{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\references{
Currently none
}
\seealso{
Currently none
}
\keyword{Currently}
\keyword{none}

