% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REYNqaqc_P5.R
\name{NOISE_rs}
\alias{NOISE_rs}
\title{Detection limit for fluxes}
\usage{
NOISE_rs(eddy.data_random = eddy.data_loc, REYN_loc = REYN_loc,
  AlgBase = AlgBase, FcorPOT = FcorPOT, FcorPOTl = FcorPOTl,
  whr_flux = whr_flux, conf_level = 0.95, crit_iter = 0.01)
}
\arguments{
\item{Currently}{none}
}
\value{
Currently none
}
\description{
Function defintion. Detection limit for fluxes.
}
\examples{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\references{
Currently none
}
\seealso{
Currently none
}
\keyword{eddy-covariance,}
\keyword{flux}
\keyword{turbulent}

