% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.r
\name{OGIV.plot}
\alias{OGIV.plot}
\title{Plot ogives}
\usage{
OGIV.plot(name, IDR, RAW, IDS, SMO, thresh, Fperi, title = NULL,
  labx = NULL, laby = NULL, colvec = c("grey", 1, 1), LIMX = NULL,
  LIMY = NULL, weight = F, wl = T, norm = 1)
}
\arguments{
\item{Currently}{none}
}
\value{
Currently none
}
\description{
Plot ogives.
}
\examples{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\references{
Currently none
}
\seealso{
Currently none
}
\keyword{Currently}
\keyword{none}

