% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIT.R
\name{PFIT_apply}
\alias{PFIT_apply}
\title{Definition function: Perform planar fit transformation}
\usage{
PFIT_apply(u_m = data.frame(xaxs, yaxs, zaxs), al = 0, be = 0, b0 = 0)
}
\arguments{
\item{Currently}{none}
}
\value{
Currently none
}
\description{
Function defintion. Performs planar fit transformation on wind vector in right-hand body coordinate system, e.g. CSAT3: Positive from front, left, below.
}
\examples{
Currently none
}
\author{
Ke Xu \email{xuke2012abroad@gmail.com},
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\references{
Currently none
}
\seealso{
Currently none
}
\keyword{coordinate}
\keyword{fit,}
\keyword{planar}
\keyword{rotation,}
\keyword{transformation}

