% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIT.R
\name{PFIT_det}
\alias{PFIT_det}
\title{Definition function: Determine planar fit coefficients}
\usage{
PFIT_det(u, v, w)
}
\arguments{
\item{Currently}{none}
}
\value{
Currently none
}
\description{
Function defintion. Determines planar fit coefficients on wind vector in right-hand body coordinate system, e.g. CSAT3: Positive from front, left, below.
}
\examples{
Currently none
}
\references{
Currently none
}
\seealso{
Currently none
}
\author{
Ke Xu \email{xuke2012abroad@gmail.com},
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\keyword{coordinate}
\keyword{fit,}
\keyword{planar}
\keyword{rotation,}
\keyword{transformation}
