% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REYNqaqc_P5.R
\name{REYNerro_FD_mole_dry}
\alias{REYNerro_FD_mole_dry}
\title{Definition function: Statistical errors for fluxes}
\usage{
REYNerro_FD_mole_dry(isca, mn, rc, Lf, COI = 0)
}
\arguments{
\item{Currently}{none}
}
\value{
Currently none
}
\description{
Function defintion. Statistical errors for fluxes.
}
\examples{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\references{
Currently none
}
\seealso{
Currently none
}
\keyword{eddy-covariance,}
\keyword{flux}
\keyword{turbulent}

