% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REYNflux_P5.R
\name{REYNflux_FD_mole_dry}
\alias{REYNflux_FD_mole_dry}
\title{Definition function: Calculate turbulent vertical flux and auxiliary variables}
\usage{
REYNflux_FD_mole_dry(data = eddy.data, AlgBase = c("mean", "trnd",
  "ord03")[1], FcorPOT = TRUE, FcorPOTl = NULL, PltfEc = "airc",
  flagCh4 = TRUE)
}
\arguments{
\item{Currently}{none}
}
\value{
Currently none
}
\description{
Function defintion. Calculate turbulent vertical flux and auxiliary variables.
}
\examples{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\references{
Currently none
}
\seealso{
Currently none
}
\keyword{eddy-covariance,}
\keyword{flux}
\keyword{turbulent}

