% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REYNqaqc_P5.R
\name{REYNitcs}
\alias{REYNitcs}
\title{Integral turbulence characteristics}
\usage{
REYNitcs(sigma, latitude, stdev, scale)
}
\arguments{
\item{Currently}{none}
}
\value{
Currently none
}
\description{
Function defintion. Integral turbulence characteristics.
}
\examples{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\references{
Currently none
}
\seealso{
Currently none
}
\keyword{eddy-covariance,}
\keyword{flux}
\keyword{turbulent}

