% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REYNqaqc_P5.R
\name{REYNstat_FD_mole_dry}
\alias{REYNstat_FD_mole_dry}
\title{Stationarity tests}
\usage{
REYNstat_FD_mole_dry(data, type = c(1, 2, 3)[2], whr_flux, NOsusa = 6,
  FcorPOT, FcorPOTl = FcorPOTl)
}
\arguments{
\item{Currently}{none}
}
\value{
Currently none
}
\description{
Function defintion. Stationarity tests.
}
\examples{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\references{
Currently none
}
\seealso{
Currently none
}
\keyword{eddy-covariance,}
\keyword{flux}
\keyword{turbulent}

