% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REYNqaqc_P5.R
\name{roughness}
\alias{roughness}
\title{Aerodynamic roughness length}
\usage{
roughness(d_z_m, d_L_v_0, u_hor, u_star, sigma_range = c(-2, 1))
}
\arguments{
\item{Currently}{none}
}
\value{
Currently none
}
\description{
Function defintion. Aerodynamic roughness length.
}
\examples{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\references{
Currently none
}
\seealso{
Currently none
}
\keyword{eddy-covariance,}
\keyword{flux}
\keyword{turbulent}

