% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.r
\name{SPEC.fwd}
\alias{SPEC.fwd}
\title{Definition function: Fast Fourier transform}
\usage{
SPEC.fwd(
  tstamp = POST$UT,
  data,
  relmot = POST$veloTAmagn,
  height = POST$lengMCzLS,
  fre = freq,
  demean = F,
  detrend = T,
  taper = 0.05
)
}
\arguments{
\item{Currently}{none}
}
\value{
Currently none
}
\description{
Fast Fourier transform.
}
\examples{
Currently none
}
\references{
Currently none
}
\seealso{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\keyword{Currently}
\keyword{none}
