% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REYNqaqc_P5.R
\name{univfun}
\alias{univfun}
\title{Integral over the universal function}
\usage{
univfun(d_z_m, d_L_v_0, sigma_range = c(-2, 1))
}
\arguments{
\item{Currently}{none}
}
\value{
Currently none
}
\description{
Function defintion. Integral over the universal function.
}
\examples{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\references{
Currently none
}
\seealso{
Currently none
}
\keyword{eddy-covariance,}
\keyword{flux}
\keyword{turbulent}

