% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.flux.R
\name{wrap.flux}
\alias{wrap.flux}
\title{Wrappter function: Calculate turbulent vertical flux and auxiliary variables}
\usage{
wrap.flux(
  data,
  AlgBase = c("mean", "trnd", "ord03")[1],
  SlctPot = FALSE,
  PresPot = eddy4R.base::IntlNatu$Pres00,
  ...
)
}
\arguments{
\item{Currently}{none}
}
\value{
Currently none
}
\description{
Function wrapper. Calculate turbulent vertical flux and auxiliary variables.
}
\examples{
Currently none
}
\references{
Currently none
}
\seealso{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\keyword{eddy-covariance,}
\keyword{flux}
\keyword{turbulent}
