% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.rot.R
\name{wrap.rot}
\alias{wrap.rot}
\title{Define Rotation of Wind Vectors}
\usage{
wrap.rot(data, mn, rotType = c("single", "double", "planarFit")[1],
  plnrFitCoef = NULL)
}
\arguments{
\item{data}{input data to REYNflux [data.frame]}

\item{mn}{mean of input data. created during "TIME SERIES AVERAGES" step [data.frame]}

\item{rotType}{type of rotation to be performed, one of "single", "double" or "planarFit" (once implemented) [character vector]}

\item{plnrFitCoef}{coefficients for planar fit [numeric vector]}
}
\value{
list containing updated data and mn objects, also B and BT rotation matricies in single or double cases
}
\description{
Perform either single or double rotation of wind vectors. TODO: include planar fit rotation based on suplied coefficients
}
\author{
W. S. Drysdale
}
