% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.spec.freq.cnst.R
\name{def.spec.freq.cnst}
\alias{def.spec.freq.cnst}
\title{Definition function: 63\% frequency constant after Aubinet (2012) Eq. 4.22}
\usage{
def.spec.freq.cnst(inpPrd)
}
\arguments{
\item{inpPrd}{numeric, time period in seconds}
}
\value{
A numeric vector \code{freqOut} of equivalent frequencies to the input time periods
}
\description{
63\% frequency constant after Aubinet (2012) Eq. 4.22.
}
\examples{

inpPrd <- seq(1, 3600, by = 10)
def.spec.freq.cnst(inpPrd = inpPrd)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
Aubinet (2012)
}
\seealso{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}

David Durden
}
\keyword{FFT,}
\keyword{Fast}
\keyword{Fourier}
\keyword{Transform,}
\keyword{spectral}
