% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.spec.freq.cut.R
\name{def.spec.freq.cut}
\alias{def.spec.freq.cut}
\title{Definition function: Determine half-power / cut-off frequency}
\usage{
def.spec.freq.cut(
  FreqCut = 0.5,
  FreqPeak,
  idep,
  depe,
  MethSpec = c("spec", "cosp")[2],
  paraStbl,
  MethWght = FALSE,
  ThshFreqRng = c(0.01, 10),
  CoefCor,
  NumBin = 1000,
  FilePlot = NULL
)
}
\arguments{
\item{FreqCut}{half-power / cut-off frequency}

\item{FreqPeak}{frequency f at which fCO(f) reaches its maximum value}

\item{idep}{independent variable, preferabley f, but n is possible}

\item{depe}{dependent variable, spectra or cospectra}

\item{MethSpec}{spectrum or cospectrum?}

\item{paraStbl}{stability parameter}

\item{MethWght}{use frequency-weighted (co)spectrum?}

\item{ThshFreqRng}{frequency range for determining optimiality criterion}

\item{CoefCor}{correction factor for spectral attenuation}

\item{NumBin}{number of bins if binning shall be performed}

\item{FilePlot}{generate plot?}
}
\value{
Optimality criterion for transfer function to correct the data to model specta
}
\description{
Determine half-power / cut-off frequency.
}
\examples{
Currently none
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
Foken T., Leuning R., Oncley S.R., Mauder M., Aubinet M. (2012) Corrections and Data Quality Control. In: Aubinet M., Vesala T., Papale D. (eds) Eddy Covariance. Springer Atmospheric Sciences. Springer, Dordrecht. https://doi.org/10.1007/978-94-007-2351-1_4
Eugster, W. and W. Senn (1995). "A COSPECTRAL CORRECTION MODEL FOR MEASUREMENT OF TURBULENT NO2 FLUX." Boundary-Layer Meteorology 74(4): 321-340.
Foken, Thomas. (2017). Micrometeorology. 10.1007/978-3-642-25440-6.
}
\seealso{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}

David Durden
}
\keyword{FFT,}
\keyword{Fast}
\keyword{Fourier}
\keyword{Transform,}
\keyword{cut-off}
\keyword{frequency,}
\keyword{half-power}
\keyword{spectral,}
