% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.spec.modl.R
\name{def.spec.modl}
\alias{def.spec.modl}
\title{Definition function: Model (co)spectrum after Massman, 2005 (in Lee, 2005)}
\usage{
def.spec.modl(
  idep,
  MethSpec = c("spec", "cosp")[2],
  paraStbl,
  FreqPeak = 0.1,
  MethWght = TRUE
)
}
\arguments{
\item{idep}{independent variable, preferabley f, but n is possible}

\item{MethSpec}{spectrum or cospectrum?}

\item{paraStbl}{stability parameter}

\item{FreqPeak}{frequency f at which fCO(f) reaches its maximum value}

\item{MethWght}{output frequency-weighted (co)spectrum?}
}
\value{
Return a (co)spectrum model after Massman, 2005 (in Lee, 2005).
}
\description{
Model (co)spectrum after Massman, 2005 (in Lee, 2005).
}
\examples{
Currently none
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
Massman, 2005 (in Lee, 2005)
Kaimal (1972)
}
\seealso{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}

David Durden
}
\keyword{FFT,}
\keyword{Fast}
\keyword{Fourier}
\keyword{Transform,}
\keyword{spectral}
