% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.spec.tfun.sigm.R
\name{def.spec.tfun.sigm}
\alias{def.spec.tfun.sigm}
\title{Definition function: Sigmoidal transfer function (Lorentzian)}
\usage{
def.spec.tfun.sigm(FreqCut, Freq)
}
\arguments{
\item{FreqCut}{half-power / cut-off frequency}

\item{Freq}{frequency}
}
\value{
A Sigmoidal transfer function for the defined frequencies and cut-off frequency.
}
\description{
Sigmoidal transfer function (Lorentzian) after Eugster and Senn (1995) in Aubinet et al. (2012) Eq. 4.21.
}
\examples{
Currently none
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
Aubinet et al. (2012)
Eugster and Senn (1995)
}
\seealso{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}

David Durden
}
\keyword{FFT,}
\keyword{Fast}
\keyword{Fourier}
\keyword{Transform,}
\keyword{spectral}
