% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.spec.time.cnst.R
\name{def.spec.time.cnst}
\alias{def.spec.time.cnst}
\title{Definition function: 63\% time constant after Aubinet (2012) Eq. 4.22}
\usage{
def.spec.time.cnst(inpFreq)
}
\arguments{
\item{inpFreq}{numeric, frequency in Hertz (Hz)}
}
\value{
A numeric vector \code{prdOut} of equivalent time periods to the input frequencies in seconds (sec)
}
\description{
63\% time constant after Aubinet (2012) Eq. 4.22.
}
\examples{

inpFreq <- seq(1, 0.02, by = -0.02)
def.spec.time.cnst(inpFreq = inpFreq)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
Aubinet (2012)
}
\seealso{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}

David Durden
}
\keyword{FFT,}
\keyword{Fast}
\keyword{Fourier}
\keyword{Transform,}
\keyword{spectral}
