% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.wave.vari.R
\name{def.vari.wave}
\alias{def.vari.wave}
\title{Definition function: function to determine the temporally resolved variance/covariance from continuous wavelet transform}
\usage{
def.vari.wave(
  spec01 = mycwt[["w_met"]]@spectrum,
  spec02 = NULL,
  scal = mycwt[[vari]]@scale,
  prd = mycwt[[vari]]@period,
  freq_0 = NA,
  whr_peri = NULL,
  coefNorm = rpt$coefNorm,
  flag,
  paraStbl,
  SC
)
}
\arguments{
\item{spec01}{Waves package output object spectrum, continuous wavelet transform output object complex spectrum for the first variable (typically w')denoted as \code{object1@spectrum}.}

\item{spec02}{Waves package output object spectrum, continuous wavelet transform output object complex spectrum for the second variable for cospectra denoted as \code{object2@spectrum}.}

\item{scal}{Waves package output object scale, width of the wavelet at each scale [s] denoted as \code{object1@scale}}

\item{prd}{Waves package output object period, approximate corresponding Fourier period [s] denoted as \code{object1@period}}

\item{freq_0}{vector, half-power frequencies for individual variables [Hz] for determining transfer function to correct frequency response}

\item{whr_peri}{numeric, which wavelengths/spatial scales to consider if you want to only consider high frequency values}

\item{coefNorm}{numeric, normalization factor specific to the choice of Wavelet parameters.}

\item{flag}{Wavelet flag: process (0) or not (1)}

\item{paraStbl}{stability parameter (numeric)}

\item{SC}{spectrum or cospectrum  c("spe", "cos")?}
}
\value{
A vector constaining temporally resolved variance/covariance from the continuous wavelet transform.
}
\description{
Definition function. Function to determine the temporally resolved variance/covariance from continuous wavelet transform including high-frequency spectral correction and selectable low-frequency cutoff. The frequency response correction using Wavelet techniques described in Norbo and Katul, 2012 (NK12)
}
\examples{
Currently none.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none.
}
\author{
David Durden \email{ddurden@battelleecology.org}
Stefan Metzger
}
\keyword{NK12,}
\keyword{Wavelet,}
\keyword{correction}
\keyword{cospectrum,}
\keyword{frequency}
\keyword{response}
\keyword{spectrum,}
