% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.r
\name{SPEC.plot}
\alias{SPEC.plot}
\title{Definition function: Plot power spectra of up to three variables}
\usage{
SPEC.plot(name, IDE, DEP, legtext, labx = NA, laby = NA, colvec = NULL,
  limx = NULL, limy = NULL, setoff = NULL, powerc = NULL,
  inisub = c(5e-02, Inf))
}
\arguments{
\item{Currently}{none}
}
\value{
Currently none
}
\description{
Plot power spectra of up to three variables.
}
\examples{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\references{
Currently none
}
\seealso{
Currently none
}
\keyword{Currently}
\keyword{none}

