% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.rot.R
\name{wrap.rot}
\alias{wrap.rot}
\title{Define Rotation of Wind Vectors}
\usage{
wrap.rot(data, rotType = c("single", "double", "planarFit", "none")[1],
  plnrFitCoef = NULL, plnrFitType = c("simple", "time", "wind")[1])
}
\arguments{
\item{data}{data.frame containing u_met, v_met and w_met [data.frame]}

\item{rotType}{type of rotation to be performed, one of "single", "double" or "planarFit" [character vector]}

\item{plnrFitCoef}{coefficients for planar fit [numeric vector or data.frame]}

\item{plnrFitType}{type of planar fit, "simple", "date" or "wind". [character vector] \itemize{
                   \item simple - numeric vector constant of coefficeients, or coefficeients that are controlled from the workflow. c(al,be,b0)
                   \item time - data.frame with columns date, al, be, b0. values with date nearest to mean(data$date)
                   \item time - data.frame with columns PSI_uv, al, be, b0. values with date nearest to average wind dir
}}
}
\value{
Data object with rotated wind vectors added
}
\description{
Perform either single or double rotation of wind vectors.
}
\author{
W. S. Drysdale
}
