% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.wave.R
\name{wrap.wave}
\alias{wrap.wave}
\title{Wrapper function: Calculate Wavelet spectrum/cospectrum and frequency response correction}
\usage{
wrap.wave(
  dfInp,
  DiffScal = 1/8,
  FuncWave = Waves::morlet(),
  FreqSamp = 20,
  ThshMiss = 0.1,
  paraStbl
)
}
\arguments{
\item{dfInp}{data.frame, consisting of the input data to perform the wavelet transformation}

\item{DiffScal}{numeric, determining the step difference in scales for the wavelet transformation}

\item{FuncWave}{Waves package function, denoting the type of mother wavelet function to be used in the transformation}

\item{FreqSamp}{numeric, that determines the time series objects points}

\item{ThshMiss}{numeric, dimensionless fraction of missing values in each column of data allowed before the quality flag is tripped. Defaults to 0.1 or 10 percent.}

\item{paraStbl}{stability parameter (numeric)}
}
\value{
An list constaining wavelet spectra, quality flags if data was available to perform correction, and frequency reponse correction parameters if activated.
}
\description{
Wrapper function. Calculate Wavelet spectrum/cospectrum using the Waves package. The frequency response correction using Wavelet techniques described in Norbo and Katul, 2012 (NK12)
}
\examples{
Currently none.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none.
}
\author{
David Durden \email{ddurden@battelleecology.org}
}
\keyword{NK12,}
\keyword{Wavelet,}
\keyword{correction}
\keyword{cospectrum,}
\keyword{frequency}
\keyword{response}
\keyword{spectrum,}
